+++ django_fido/tests/test_models.py
@@ -179,7 +179,16 @@ class TestAuthenticatorMetadata(TestCase):
         self.assertEqual(metadata.level, AuthLevel.L0)
 
     def test_level_revoked(self):
-        status = {'statusReports': [{'status': 'FIDO_CERTIFIED'}, {'status': 'REVOKED'}]}
+        status = {'statusReports': [{'status': 'FIDO_CERTIFIED', 'effectiveDate': '2021-03-05'},
+                                    {'status': 'REVOKED', 'effectiveDate': '2021-03-06'}]}
+        metadata = AuthenticatorMetadata.objects.create(identifier='95442b2e-f15e-4def-b270-efb106facb4e',
+                                                        metadata_entry=json.dumps(status))
+        self.assertEqual(metadata.level, AuthLevel.NONE)
+
+    def test_level_revoked_reversed(self):
+        status = {'statusReports': [{'status': 'REVOKED', 'effectiveDate': '2021-03-06'},
+                                    {'status': 'FIDO_CERTIFIED', 'effectiveDate': '2021-03-05'}]}
+
         metadata = AuthenticatorMetadata.objects.create(identifier='95442b2e-f15e-4def-b270-efb106facb4e',
                                                         metadata_entry=json.dumps(status))
         self.assertEqual(metadata.level, AuthLevel.NONE)
@@ -190,9 +199,17 @@ class TestAuthenticatorMetadata(TestCase):
         self.assertEqual(metadata.level, AuthLevel.NONE)
 
     def test_level_breach_fixed(self):
-        status = {'statusReports': [{'status': 'FIDO_CERTIFIED'},
-                                    {'status': 'USER_VERIFICATION_BYPASS'},
-                                    {'status': 'UPDATE_AVAILABLE'}]}
+        status = {'statusReports': [{'status': 'FIDO_CERTIFIED', 'effectiveDate': '2021-03-05'},
+                                    {'status': 'USER_VERIFICATION_BYPASS', 'effectiveDate': '2021-03-06'},
+                                    {'status': 'UPDATE_AVAILABLE', 'effectiveDate': '2021-03-07'}]}
+        metadata = AuthenticatorMetadata.objects.create(identifier='95442b2e-f15e-4def-b270-efb106facb4e',
+                                                        metadata_entry=json.dumps(status))
+        self.assertEqual(metadata.level, AuthLevel.L0)
+
+    def test_level_breach_fixed_unordered(self):
+        status = {'statusReports': [{'status': 'UPDATE_AVAILABLE'},
+                                    {'status': 'USER_VERIFICATION_BYPASS', 'effectiveDate': '2021-03-06'},
+                                    {'status': 'FIDO_CERTIFIED', 'effectiveDate': '2021-03-05'}]}
         metadata = AuthenticatorMetadata.objects.create(identifier='95442b2e-f15e-4def-b270-efb106facb4e',
                                                         metadata_entry=json.dumps(status))
         self.assertEqual(metadata.level, AuthLevel.L0)